Video Augmentation
I was looking for a good video augmentation library that I could use in Anaconda without any problems and I did'nt find any, so I picked up parts of a sort of other libraries and put it together in code to create a new, more simplified, library. I hope it help someone.